[workflow] Setup GitHub actions for CI (#10)
This pull request setups up GitHub actions for cue android. It will build every commit in pull request.
Right now it writes to gradle.properties a dummy key for CI purposes. Eventually we want to setup this properly so everything can still be automated without leaking anything.
This is a temporary solution. Eventually the gradle.properties file that contains the secret should be properly removed from git so that we won't leak any secret.